% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{combined}
\alias{combined}
\title{The Wanderlust dataset, combined basal and IL7 conditions}
\format{A tibble of 10,000 cells by 51 features, including all the input
markers, Wanderlust values, and the condition. The first 5000 rows
are untreated cells and the last 5000 rows are IL7 treated.}
\usage{
combined
}
\description{
A single patient pair of basal and IL7 treated cells from bone marrow
gated for B cell precursors.
}
\keyword{datasets}
