% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.processing.R
\name{data.to.fcs}
\alias{data.to.fcs}
\title{Takes as input a labeled data matrix,
and outputs an fcs file containing said inroamtion.}
\usage{
data.to.fcs(dta, outfile, untransform)
}
\arguments{
\item{outfile}{a string containing your original data matrix}

\item{untransform}{takes data out of asinh(x/5) transformation}

\item{dat}{a data matrix intended to be converted to fcs}
}
\value{
fcs file containing your original data matrix
}
\description{
Note that Cytobank's DROP feature in theory does what
this script does for csv files, but its availability is currently limited
and not all fcs processing software will have similar functionality.
Note also the untransform parameter. This is for the instrances where
an asinh transformation with a scale argument of 5 takes place within R
(for example when one is clustering). Cytobank is not used to reading these
scales. Thus, untransforming them prior to Cytobank use allows the biaxials
to be effectively visualized.
}
\examples{
basal <- system.file('extdata',
    'Bendall et al Cell Sample C_basal.fcs',
    package = "Sconify")
basal <- fcs.to.tibble(basal, transform = "asinh")
data.to.fcs(basal, "basal.output.FCS")
}
