% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{final}
\alias{final}
\title{Post-scone output of the "combiend" Wanderlust data.}
\format{A tibble of 10,000 cells and 87 feaures, including the input
features, the SCONE-generated comparisons, differential abundance, and
two t-SNE dimesnions}
\usage{
final
}
\description{
"combined" data taken through KNN generation and comparisons, along
with t-SNE map generation.
}
\keyword{datasets}
