% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{final.sub}
\alias{final.sub}
\title{A 2000 cell subset of the "final" data}
\format{A tibble of 2000 cells and 87 feaures, including the input
features, the SCONE-generated comparisons, differential abundance, and two
t-SNE runs. One genrated from the 10,000 cell data (bh-SNE1 and bh-SNE2),
along with a run from the 2000 cell data (bh-SNE11 and bh-SNE21). This
is used in the post-processing vignette to describe the subsample.and.tsne
function from this package}
\usage{
final.sub
}
\description{
2000 cells randomly taken from the 10,000 cell final dataset.
}
\keyword{datasets}
