% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.and.statistics.R
\name{fnn}
\alias{fnn}
\title{Compute knn using the fnn algorithm}
\usage{
fnn(cell.df, input.markers, k = 100)
}
\arguments{
\item{cell.df}{the cell data frame used as input}

\item{input.markers}{markers to be used as input for knn}

\item{k}{the number of nearest neighbors to identify}
}
\value{
nn: list of 2, nn.index: index of knn (columns) for each cell (rows)
nn.dist: euclidean distance of each k-nearest neighbor
}
\description{
This function is a wrapper around FNN package
functionality to speed up the KNN process. It uses KD trees as default,
along with k set to 100. Selection of K will vary based on the dataset.
See k.selection.R.
}
\examples{
fnn(combined, input.markers)
}
