% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{funct.markers}
\alias{funct.markers}
\title{Functional markers from the Wanderlust dataset.}
\format{
A vector of strings.
}
\usage{
funct.markers
}
\description{
These are the markers that will be used in the KNN comparisons, as opposed
to the KNN generation.
}
\keyword{datasets}
