% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.and.statistics.R
\name{get.knn.de}
\alias{get.knn.de}
\title{Get the KNN density estimaion}
\usage{
get.knn.de(nn.matrix)
}
\arguments{
\item{nn.matrix}{A list of 2, where the first is a matrix of nn indices,
and the second is a matrix of nn distances}
}
\value{
a vector where each element is the KNN-DE for that given cell,
ordered by row number, in the original input matrix of cells x features
}
\description{
Obtain a density estimation derived from the original manifold,
avoiding the lossiness of lower dimensional embeddings
}
\examples{
ex.knn <- fnn(wand.combined, input.markers)
get.knn.de(ex.knn)
}
