% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.and.statistics.R
\name{GetKnnDe}
\alias{GetKnnDe}
\title{Get the KNN density estimaion}
\usage{
GetKnnDe(nn.matrix)
}
\arguments{
\item{nn.matrix}{A list of 2, where the first is a matrix of nn indices,
and the second is a matrix of nn distances}
}
\value{
a vector where each element is the KNN-DE for that given cell,
ordered by row number, in the original input matrix of cells x features
}
\description{
Obtain a density estimation derived from the original manifold,
avoiding the lossiness of lower dimensional embeddings
}
\examples{
ex.knn <- Fnn(wand.combined, input.markers, k = 30)
GetKnnDe(ex.knn)
}
