% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{il7.data}
\alias{il7.data}
\title{IL7 data}
\format{A tibble of 10,000 cells by 51 features. All markers in the
dataset, along with pre-calculated Wanderlust value and condition,
which is a string that denotes that this is the "IL7" condition for
each row. Important when this is concatenated with additional conditions}
\usage{
il7.data
}
\description{
The IL7 treated cells from a single patient in the Wanderlust dataset
}
\keyword{datasets}
