% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k.selection.R
\name{impute}
\alias{impute}
\title{Imputes values for all markers (used as input) for each cell}
\usage{
impute(cells, input.markers, nn)
}
\arguments{
\item{cells}{the input matrix of cells}

\item{input.markers}{the markers the user wants to impute}

\item{nn}{the matrix of k-nearest neighbors
(derived perhaps NOT from the "input markers" above)}
}
\value{
a data frame of imputed cells for the "input markers" of interest
}
\description{
This function takes as input the markers to be imputed from
a pre-existing KNN computation.
}
