% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k.selection.R
\name{impute.testing}
\alias{impute.testing}
\title{Impute testing}
\usage{
impute.testing(k.titration, cells, input.markers, test.markers)
}
\arguments{
\item{k.titration}{a vector integer values of k to be tested}

\item{cells}{a matrix of cells by features used as original input}

\item{input.markers}{markers to be used for the knn calculation}

\item{test.markers}{the markers to be tested for imputation (either surface or scone)}
}
\value{
the median imputation error for each value k tested
}
\description{
Tests the euclidean distance error for imputation using knn and markers of interest
}
\examples{
impute.testing(k.titration = c(10, 20),
  cells = wand.combined,
  input.markers = input.markers,
  test.markers = funct.markers)
}
