% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{input.markers}
\alias{input.markers}
\title{Input markers for the Wanderlust dataset}
\format{
A vector of strings corresponding to the markers.
}
\usage{
input.markers
}
\description{
These are the markers that KNN generation will be done on for the
Wanderlust dataset. These are mostly surface markers.These are the
same markers one would use as input for clustering or t-SNE generation,
for exmaple, as they are not expected to change through the duration
of the quick IL7 stimulation.
}
\keyword{datasets}
