% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{k.titration}
\alias{k.titration}
\title{K titration}
\format{A vector of numbers}
\usage{
k.titration
}
\description{
A titration of values of K used for a 1000 cell dataset as input
for the impute.testing function.
}
\keyword{datasets}
