% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{knn.density}
\alias{knn.density}
\title{KNN density}
\format{A numeric vector}
\usage{
knn.density
}
\description{
A vector of per-cell KNN density from the Wanderlust dataset, generated
from the get.knn.de command. Of note, the SCONE pipeline will automatically
generate this for you. This is just one divided by the average distance
to KNN.
}
\keyword{datasets}
