% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{knn.list}
\alias{knn.list}
\title{KNN list}
\format{A list of tibbles, with each element being a tibble of cells
by features corresponding to a given cell's KNN. The element of the list
corresponds to the cell identity (row) of the original input data.}
\usage{
knn.list
}
\description{
A list of the features of each cell's KNN. This is from the optional
make.knn.list function, that allows the user to do KNN-based operations
that go beyond the scope of the Sconify package.
}
\keyword{datasets}
