% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.processing.R
\name{LogTransformQ}
\alias{LogTransformQ}
\title{Log transform the q values}
\usage{
LogTransformQ(dat, negative)
}
\arguments{
\item{dat}{tibble containing cells x features, with orignal expression,
p values, and raw change}

\item{negative}{boolean value to determine whether to multiple transformed
p values by -1}
}
\value{
result: tibble of cells x features with all p values log10
transformed
}
\description{
Takes all p values from the data and does a log10 transform
for easier visualization.
}
