% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.processing.R
\name{make.hist}
\alias{make.hist}
\title{make.hist}
\usage{
make.hist(dat, k, column.label, x.label)
}
\arguments{
\item{dat}{tibble consisting both of original markers and the appended
values from scone}

\item{k}{the binwidth, set to 1/k}

\item{column.label}{the label in the tibble's columns the function will
search for}

\item{x.label}{the label that the x axis will be labeled as}
}
\value{
a histogram of said vector in ggplot2 form
}
\description{
Makes a histogram of the data that is inputted
}
\examples{
make.hist(wand.final, 100, "IL7.fraction.cond.2", "fraction IL7")
}
