% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{md.final}
\alias{md.final}
\title{Multiple donor dataset post-SCONE}
\format{A tibble of 18,514 cells by 124 features. These features include
original input markers, per-donor comparisons, per-cell comparisons,
density, density estimation, and per-condition differential abundance.}
\usage{
md.final
}
\description{
The Fragidakis dataset with multiple donors, post-SCONE analysis.
}
\keyword{datasets}
