% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{md.input}
\alias{md.input}
\title{Multiple donor dataset input markers}
\format{A vector of strings}
\usage{
md.input
}
\description{
The markers to be used for KNN generation for the Fragidakis multiple donor
dataset. These are primarily surface markers
}
\keyword{datasets}
