% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{md.nn}
\alias{md.nn}
\title{Fragidakis multiple donor dataset nearest neighbor information}
\format{List of 2. First element named nn.index is a matrix of 19,992 cells
by 200 nearest neighbor positions, with the matrix elements being cell
identity.The second element is a matrix of 19,992 cells by 200 nearest
neighborhood positions with the matrix elements being distances.}
\usage{
md.nn
}
\description{
Two matrices corresponding to the KNN identity and distance.
}
\keyword{datasets}
