% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{md}
\alias{md}
\title{The Fragidakis multiple donor dataset.}
\format{A tibble of 19,992 cells by 47 features. These features include
all acquired markers, the per-cell condition (baseline, LPS, or IL6), and
the donor (coded p101, p104, p106, and p108).}
\usage{
md
}
\description{
The two condition tibble used as input for SCONE analysis.
}
\keyword{datasets}
