% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{md.scone.output}
\alias{md.scone.output}
\title{The SCONE output of the Fragidakis multiple donor dataset.}
\format{A tibble of 19,992 cells and 75 features, that includes the
aforementioned features, but not the original input features. These are
added in the post-processing step along with a t-SNE map.}
\usage{
md.scone.output
}
\description{
The KNN comparisons, density, and per-cell differential abundance from the
aforementioned dataset.
}
\keyword{datasets}
