% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{md.scone}
\alias{md.scone}
\title{KNN comparison markers for Fragidakis multiple donor dataset.}
\format{A vector of strings}
\usage{
md.scone
}
\description{
The markers in the Fragidakis multiple donor dataset to be used for KNN
(scone) comparisons
}
\keyword{datasets}
