% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.files.R
\name{meaning.of.life}
\alias{meaning.of.life}
\title{Meaning of life}
\usage{
meaning.of.life()
}
\value{
A string containing a random musing
}
\description{
Just a random musing
}
\examples{
meaning.of.life()
}
