% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.and.statistics.R
\name{MultipleDonorStatistics}
\alias{MultipleDonorStatistics}
\title{Runs a t test on the medians or means of multiple donors for the
same condition}
\usage{
MultipleDonorStatistics(basal, stim, stim.name, donors)
}
\arguments{
\item{basal}{tibble that contains unstim for a knn including donor identity}

\item{stim}{tibble that contains stim for a knn including donor identity}

\item{stim.name}{string of the name of the current stim being tested}

\item{donors}{vector of strings corresponding to the designated names
of the donors}
}
\value{
result: a named vector of p values (soon to be q values) from the
t test done on each marker
}
\description{
This function is for the instance that multiple donors are
being compared against each other within the k-nearest neighborhood of
interest. The mean value of the markers of interest is calculated across
the donors, such that each data point for the subsequent t-test represents
a marker for a danor.
}
