% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nn}
\alias{nn}
\title{Nearest neighbors}
\format{List of 2. First element named nn.index is a matrix of 10,000 cells
by 100 nearest neighbor positions, with the matrix elements being cell
identity.The second element is a matrix of 10,000 cells by 100 nearest
neighborhood positions with the matrix elements being distances.}
\usage{
nn
}
\description{
Nearest neighbor matrices for the Wanderlust dataset.
}
\keyword{datasets}
