% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.files.R
\name{ParseMarkers}
\alias{ParseMarkers}
\title{Parse markers contained in a Sconify-directed marker file}
\usage{
ParseMarkers(marker.file)
}
\arguments{
\item{marker.file}{modified markers.csv file, now containing two columns.
the left column containing KNN input markers, and the right column
containing KNN comparison markers}
}
\value{
a list of 2 vectors of strings. The first element, labeled "input"
is a vector KNN input markers. THe second slemenet, labeled "functional"
are the markers to be used in the KNN based comparisons
}
\description{
This occurs after the user has modified the markers.csv file
to determine which markers are to be used as input for KNN and which markers
are to be used for within-knn comparisons
}
\examples{
file <- system.file("extdata", "Bendall_et_al_Cell_Sample_C_basal.fcs",
    package = "Sconify")
GetMarkerNames(file)
}
