% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.files.R
\name{process.multiple.files}
\alias{process.multiple.files}
\title{Converts multiple files into a concatenated data frame}
\usage{
process.multiple.files(files, transform = "asinh", numcells = 10000,
  norm = FALSE, scale = FALSE, input, name.multiple.donors = FALSE)
}
\arguments{
\item{files:}{a vector of file names (name = "anything_condition.fcs")}

\item{numcells:}{desiered number of cells in the matrix}

\item{norm:}{boolean that quantile normalizes the data if true}

\item{scale:}{boolean that converts all values to z scores if true}

\item{name.multiple.donors:}{boolean indicating whether multiple donors
will be distinguished (as a separate "patient" column)}
}
\value{
result: a combined file set
}
\description{
This is tailored to a very specific file format for unstim/stim
Files need the following name convention: "xxxx_stim.fcs"
Files where you want to name the patients need the following convention:
"xxxx__patientID_stim.fcs"
}
\examples{
basal <- system.file('extdata',
    'Bendall et al Cell Sample C_basal.fcs',
     package = "Sconify")
il7 <- system.file('extdata',
    'Bendall et al Cell Sample C_IL7.fcs',
    package = "Sconify")
markers <- system.file('extdata',
    'markers.csv',
    package = "Sconify")
markers <- read.csv(markers, stringsAsFactors = FALSE)
surface <- markers$surface
combined <- process.multiple.files(files = c(basal, il7), input = surface)
}
