% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.and.statistics.R
\name{q.correction.thresholding}
\alias{q.correction.thresholding}
\title{Corrects all p values for multiple hypotheses, sets threshold for
which change values should be reported}
\usage{
q.correction.thresholding(cells, threshold)
}
\arguments{
\item{cells:}{tibble of change values, p values, and fraction condition 2}

\item{threshold:}{a q value below which the change values will be reported
for that cell for that param. If no change is desired, this is set to 1.}
}
\value{
inputted p values, adjusted and therefore described as "q values"
}
\description{
Given the number of comparisons we make across k-nearest
neighborhoods, which is far more than that of disjoint subsetting, this
step is important given that there is an increased likelihood that some
statistically significant differences will occur by chance.
}
