% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.files.R
\name{quant.normalize.elements}
\alias{quant.normalize.elements}
\title{Takes a list of tibbles as input, and performs per-column
quantile normalization, then outputs the quantile normalized list}
\usage{
quant.normalize.elements(dat.list)
}
\arguments{
\item{dat.list:}{a list of tibbles}
}
\value{
the per-column quantile normalized list
}
