% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scone.output}
\alias{scone.output}
\title{Scone output}
\format{A tibble of 10,000 cells by 34 features. These features include
the KNN comparisons, KNN density estimation, and differential abundance.
Note that this tibble gets concatenated with the original tibble, as well
as two t-SNE dimensions in the post.processing() command of the pipeline.}
\usage{
scone.output
}
\description{
The scone output for the Wanderlust dataset
}
\keyword{datasets}
