% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wand.basal}
\alias{wand.basal}
\title{Wanderlust basal data}
\format{A tibble of 1000 cells by 51 features. All markers in the
dataset, along with pre-calculated Wanderlust value and condition,
which is a string that denotes that this is the "basal" condition for
each row. Important when this is concatenated with additional conditions}
\usage{
wand.basal
}
\description{
The basal cells from a single patient in the Wanderlust dataset
}
\keyword{datasets}
