% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wand.final}
\alias{wand.final}
\title{Post-scone output of the "combiend" Wanderlust data.}
\format{
A tibble of 1000 cells and 87 feaures, including the input
features, the SCONE-generated comparisons, differential abundance, and
two t-SNE dimesnions
}
\usage{
wand.final
}
\description{
"combined" data taken through KNN generation and comparisons, along
with t-SNE map generation.
}
\keyword{datasets}
