% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wand.il7}
\alias{wand.il7}
\title{Wanderlust IL7 data}
\format{
A tibble of 1000 cells by 51 features. All markers in the
dataset, along with pre-calculated Wanderlust value and condition,
which is a string that denotes that this is the "IL7" condition for
each row. Important when this is concatenated with additional conditions
}
\usage{
wand.il7
}
\description{
The IL7 treated cells from a single patient in the Wanderlust dataset
}
\keyword{datasets}
