% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bz.gmcsf.final.norm.scale}
\alias{bz.gmcsf.final.norm.scale}
\title{Bodenmiller-Zunder GM-CSF post-SCONE final data, that's been quantile
normalized and z scored}
\format{A tibble of 10,000 cells by 69 features. This includes all the
original parameters, the KNN-generated comparisons, differential
abundance ("fraction.cond.2), and two t-SNE coordinates.}
\usage{
bz.gmcsf.final.norm.scale
}
\description{
The post-SCONE output from a normalized and z scored Bodenmiller-Zunder
dataset pair of fcs files, one untreated and one treated with GM-CSF.
}
\keyword{datasets}
