% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bz.gmcsf.final}
\alias{bz.gmcsf.final}
\title{Bodenmiller-Zunder GM-CSF post-SCONE final data}
\format{A tibble of 9,998 cells by 69 features. This includes all the
original parameters, the KNN-generated comparisons, differential
abundance ("fraction.cond.2), and two t-SNE coordinates.}
\usage{
bz.gmcsf.final
}
\description{
The post-SCONE output from Bodenmiller-Zunder dataset pair of fcs files,
one untreated and one treated with GM-CSF.
}
\keyword{datasets}
