% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.files.R
\name{fcs.to.tibble}
\alias{fcs.to.tibble}
\title{Takes a file as input and returns a data frame of cells by features}
\usage{
fcs.to.tibble(file, transform = "asinh")
}
\arguments{
\item{file}{the fcs file containing cell infomration}

\item{transform}{if set to asinh, then asinh transforms with scale arg 5}
}
\value{
tibble of info contained within the fcs file
}
\examples{
basal <- system.file('extdata',
    'Bendall et al Cell Sample C_basal.fcs',
    package = "fcsProcess")
fcs.to.tibble(basal, transform = "none")
fcs.to.tibble(basal, transform = "asinh")
}
