% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{markers}
\alias{markers}
\title{Markers for the Wanderlust dataset}
\format{a tibble with two columns, "surface" and "fucntional."}
\usage{
markers
}
\description{
Both the surface and functional markers for the Wanderlust dataset
}
\keyword{datasets}
