% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wand.combined}
\alias{wand.combined}
\title{Wanderlust data combined basal and IL7 cells}
\format{A tibble of 1000 cells by 51 features, including all the input
markers, Wanderlust values, and the condition. The first 500 rows
are untreated cells and the last 500 rows are IL7 treated.}
\usage{
wand.combined
}
\description{
A single patient pair of basal and IL7 treated cells from bone marrow
gated for B cell precursors.
}
\keyword{datasets}
