% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wand.ideal.k}
\alias{wand.ideal.k}
\title{A named vector to help the user determine the ideal k
for the Wanderlust dataset.}
\format{A named vector, where the elements are averge imputation error
and the names are the values of from a 10,000 cell dataset.}
\usage{
wand.ideal.k
}
\description{
This is the output of the impute.testing function used on the
Wanderlust dataset, which finds the avergae imputation error of all signal
markers imputed from KNN of surface markers.
}
\keyword{datasets}
